\name{add-methods}
\docType{methods}
\alias{add}
\alias{add-methods}
\alias{add,protobufMessage-method}
\title{add elements of a repeated field of a message}
\description{
add elements to a repeated field of a message. 
}
\section{Methods}{
\describe{
\item{\code{signature(object = "protobufMessage")}}{add elements to a repeated field of a message}
}}
\keyword{methods}

