\name{ArrayInputStream-methods}
\docType{methods}
\alias{ArrayInputStream}
\alias{ArrayInputStream-methods}
\alias{ArrayInputStream,raw,missing-method}
\alias{ArrayInputStream,raw,integer-method}
\alias{ArrayInputStream,raw,numeric-method}
\title{Creates an ArrayInputStream}
\description{
Constructor for \linkS4class{ArrayInputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(payload = "raw", block_size = "missing" )}}{Creates a \linkS4class{ArrayInputStream} using the raw vector as the payload of the stream}
\item{\code{signature(payload = "raw", block_size = "integer" )}}{Creates a \linkS4class{ArrayInputStream} ... same with block size. }
\item{\code{signature(payload = "raw", block_size = "numeric" )}}{Creates a \linkS4class{ArrayInputStream} ... same with block size. }
}}
\keyword{methods}

