\name{ArrayOutputStream-methods}
\docType{methods}
\alias{ArrayOutputStream}
\alias{ArrayOutputStream-methods}
\alias{ArrayOutputStream,integer,missing-method}
\alias{ArrayOutputStream,integer,integer-method}
\alias{ArrayOutputStream,integer,numeric-method}
\alias{ArrayOutputStream,numeric,missing-method}
\alias{ArrayOutputStream,numeric,integer-method}
\alias{ArrayOutputStream,numeric,numeric-method}
\title{Creates an ArrayOutputStream}
\description{
Constructor for \linkS4class{ArrayOutputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(size = "integer", block_size = "missing" )}}{Creates a \linkS4class{ArrayOutputStream} using of the given size}
\item{\code{signature(size = "integer", block_size = "integer" )}}{Creates a \linkS4class{ArrayOutputStream} ... same with block size. }
\item{\code{signature(size = "integer", block_size = "numeric" )}}{Creates a \linkS4class{ArrayOutputStream} ... same with block size. }
\item{\code{signature(size = "numeric", block_size = "missing" )}}{Creates a \linkS4class{ArrayOutputStream} using of the given size}
\item{\code{signature(size = "numeric", block_size = "integer" )}}{Creates a \linkS4class{ArrayOutputStream} ... same with block size. }
\item{\code{signature(size = "numeric", block_size = "numeric" )}}{Creates a \linkS4class{ArrayOutputStream} ... same with block size. }
}}
\keyword{methods}

