\name{ByteCount-methods}
\docType{methods}
\alias{ByteCount}
\alias{ByteCount-methods}
\title{The number of bytes read/written since the object was created}
\description{
The number of bytes read/written since the object was created
}
\seealso{
	\linkS4class{ZeroCopyInputStream} implements ByteCount.
}
\keyword{methods}

