\name{CodedInputStream-class}
\Rdversion{1.1}
\docType{class}
\alias{CodedInputStream-class}

\title{Class "CodedInputStream" }
\description{CodedInputStream}
\section{Objects from the Class}{
TODO
}
\section{Slots}{
  \describe{
  	  \item{\code{pointer}:}{External pointer to the \code{google::protobuf::CodedInputStream}}
  }
}
\author{ Romain Francois <francoisromain@free.fr> }
\references{
	The S4 class \code{CodedInputStream} shadows the C++ class 
	CodedInputStream from the protobuf library. 
	\url{http://code.google.com/apis/protocolbuffers/docs/reference/cpp/google.protobuf.io.coded_stream.html#CodedInputStream}
}
\keyword{classes}

