\name{ConnectionInputStream-methods}
\docType{methods}
\alias{ConnectionInputStream}
\alias{ConnectionInputStream-methods}
\alias{ConnectionInputStream,connection-method}
\title{Creates an ConnectionInputStream}
\description{
Constructor for \linkS4class{ConnectionInputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(object="connection")}}{Creates a \linkS4class{ConnectionInputStream} reading from the given R binary connection.}
}}
\keyword{methods}

