\name{ConnectionOutputStream-methods}
\docType{methods}
\alias{ConnectionOutputStream}
\alias{ConnectionOutputStream-methods}
\alias{ConnectionOutputStream,connection-method}
\title{Creates an ConnectionOutputStream}
\description{
Constructor for \linkS4class{ConnectionOutputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(object="connection")}}{
	Creates a \linkS4class{ConnectionOutputStream} writing to the given R binary connection.}
}}
\keyword{methods}

