\name{descriptor-methods}
\docType{methods}

\alias{descriptor}
\alias{descriptor-methods}
\alias{descriptor,protobufMessage-method}
\title{Get the descriptor of a message}
\description{
	Get the \linkS4class{protobufDescriptor} associated
	with a \linkS4class{protobufMessage}
}
\section{Methods}{
\describe{
\item{\code{signature(object = "protobufMessage")}}{ Get the descriptor of the message, as a \linkS4class{protobufDescriptor} instance }
}}
\keyword{methods}

