\name{enum_type-methods}
\alias{enum_type}
\alias{enum_type-methods}
\title{Extract an enum type descriptor for a nested type}
\description{
	Extract a \linkS4class{EnumDescriptor} contained in a 
	\linkS4class{Descriptor}
}
\seealso{
The method is implemented for the \linkS4class{Descriptor} class
}
\keyword{methods}

