\name{fetch-methods}
\docType{methods}
\alias{fetch}
\alias{fetch-methods}
\alias{fetch,protobufMessage-method}
\title{Fetch content of a repeated field}
\description{
Fetch content of a repeated field of a message
}
\section{Methods}{
\describe{
\item{\code{signature(object = "protobufMessage")}}{Fetch content of a message repeated field}
}}
\keyword{methods}

