\name{field-methods}
\alias{field}
\alias{field-methods}
\title{Extract a field descriptor}
\description{
	Extract a \linkS4class{protobufFieldDescriptor} from a 
	\linkS4class{protobufDescriptor}
}
\seealso{
The method is implemented for the \linkS4class{protobufDescriptor} class
}
\keyword{methods}

