\name{fileDescriptor-methods}
\docType{methods}
\alias{fileDescriptor}
\alias{fileDescriptor-methods}
\alias{fileDescriptor,Descriptor-method}
\alias{fileDescriptor,Message-method}
\alias{fileDescriptor,EnumDescriptor-method}
\alias{fileDescriptor,FieldDescriptor-method}
\alias{fileDescriptor,ServiceDescriptor-method}
\alias{fileDescriptor,MethodDescriptor-method}

\title{gets the file descriptor of an object}
\description{
Gets the file descriptor of an object
}
\section{Methods}{
\describe{
\item{\code{signature(object = "Descriptor")}}{ retrieves the file descriptor associated with this descriptor}
\item{\code{signature(object = "Message")}}{ retrieves the file descriptor associated with the descriptor of this message }
\item{\code{signature(object = "EnumDescriptor")}}{ retrieves the file descriptor associated with the enum descriptor }
\item{\code{signature(object = "FieldDescriptor")}}{ retrieves the file descriptor associated with the field descriptor }
\item{\code{signature(object = "ServiceDescriptor")}}{ retrieves the file descriptor associated with the service descriptor }
\item{\code{signature(object = "MethodDescriptor")}}{ retrieves the file descriptor associated with the method descriptor }
}}
\keyword{methods}

