\name{FileInputStream-methods}
\docType{methods}
\alias{FileInputStream}
\alias{FileInputStream-methods}
\alias{FileInputStream,character,integer,logical-method}
\title{Creates an FileInputStream}
\description{
Constructor for \linkS4class{FileInputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(filename = "character", block_size = "logical", close.on.delete = "logical" )}}{Creates a \linkS4class{FileInputStream} reading from the given file.}
}}
\keyword{methods}

