\name{FileOutputStream-methods}
\docType{methods}
\alias{FileOutputStream}
\alias{FileOutputStream-methods}
\alias{FileOutputStream,character,integer,logical-method}
\title{Creates an FileOutputStream}
\description{
Constructor for \linkS4class{FileOutputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(filename = "character", block_size = "logical", close.on.delete = "logical" )}}{Creates a \linkS4class{FileOutputStream} writing to the given file.}
}}
\keyword{methods}

