\name{GetErrno-methods}
\docType{methods}
\alias{GetErrno}
\alias{GetErrno-methods}
\title{Get the error number for an I/O error}
\description{
If an I/O error has occurred on this file descriptor, 
this is the errno from that error
}
\section{Methods}{
	See classes \linkS4class{FileInputStream} and 
	\linkS4class{FileOutputStream} for implementations.
}
\keyword{methods}

