\name{invokeLocally-methods}
\docType{methods}
\alias{invokeLocally}
\alias{invokeLocally-methods}
\alias{invokeLocally,protobufMethodDescriptor,protobufMessage-method}
\title{invoke a protobuf rpc method locally}
\description{
invoke a protobuf rpc method locally
}
\section{Methods}{
\describe{
\item{\code{signature(method = "protobufMethodDescriptor", message = "protobufMessage")}}{invoke a protobuf rpc method locally}
}}
\keyword{methods}
