\name{label-methods}

\alias{label}
\alias{label-methods}

\alias{LABEL_OPTIONAL}
\alias{LABEL_REQUIRED}
\alias{LABEL_REPEATED}

\title{Gets the label of a field}
\description{
Gets the label of a field
}
\seealso{
The method is implemented for the \linkS4class{protobufFieldDescriptor} class
}
\keyword{methods}

