\name{name}
\docType{methods}
\alias{name}
\alias{name-methods}
\alias{name,Descriptor-method}
\alias{name,FieldDescriptor-method}
\alias{name,EnumDescriptor-method}
\alias{name,ServiceDescriptor-method}
\alias{name,MethodDescriptor-method}

\title{Name or full name of a descriptor}
\description{
name or full name of a descriptor
}
\section{Methods}{
\describe{
	\item{\code{signature(object = "Descriptor")}}{ ... }
	\item{\code{signature(object = "FieldDescriptor")}}{ ... }
	\item{\code{signature(object = "EnumDescriptor")}}{ ... }
	\item{\code{signature(object = "ServiceDescriptor")}}{ ... }
	\item{\code{signature(object = "MethodDescriptor")}}{ ... }
}}
\keyword{methods}

