\name{protobufFileDescriptor-class}
\Rdversion{1.1}
\docType{class}
\alias{protobufFileDescriptor-class}

\title{Class "protobufFileDescriptor"}
\description{Class "protobufFileDescriptor"}
\section{Objects from the Class}{
Objects are usually created using the \code{\link{fileDescriptor}} method
}
\section{Slots}{
  \describe{
  	  \item{\code{pointer}:}{external pointer to a \code{google::protobuf::FileDescriptor} C++ object }
  }
}
\section{Methods}{
No methods defined with class "protobufFileDescriptor" in the signature.
}
\references{ The \url{http://code.google.com/apis/protocolbuffers/docs/reference/cpp/google.protobuf.descriptor.html#FileDescriptor} }
\author{ Romain Francois <francoisromain@free.fr> }
\keyword{classes}

