\name{readASCII-methods}
\docType{methods}
\alias{readASCII}
\alias{readASCII-methods}
\alias{readASCII,Descriptor,ANY-method}
\alias{readASCII,Descriptor,character-method}
\title{read a message in ASCII format}
\description{
 Method to read a Message in ASCII format
}
\section{Methods}{
\describe{

\item{\code{signature(descriptor = "Descriptor", input = "ANY")}}{
Read the message from a connection (file, etc ...)
}

\item{\code{signature(descriptor = "Descriptor", input = "character")}}{
Read the message directly from the character string
}
}}
\keyword{methods}

