\name{set-methods}
\docType{methods}
\alias{set}
\alias{set-methods}
\alias{set,Message-method}
\title{set a subset of values of a repeated field of a message}
\description{
set a subset of values of a repeated field of a message
}
\section{Methods}{
\describe{
\item{\code{signature(object = "Message")}}{set a subset of values of a repeated field of a message}
}}
\keyword{methods}

