\name{size-methods}
\docType{methods}
\alias{size}
\alias{size-methods}
\alias{size,protobufMessage-method}
\title{Size of a message field}
\description{
The number of object currently in a given field of a protocol buffer
message. For non repeated field, it is allways 1.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "protobufMessage")}}{Number of objects in a message field}
}}
\keyword{methods}

