\name{sizegets}
\docType{methods}
\alias{size<-}
\alias{size<--methods}
\alias{size<-,Message-method}
\title{Set the size of a field}
\description{
	Sets the size of a repeated field.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "Message")}}{ sets the size of a message field }
}}
\keyword{methods}

