\name{swap-methods}
\docType{methods}
\alias{swap}
\alias{swap-methods}
\alias{swap,Message-method}
\title{swap elements of a repeated field of a message}
\description{
swap elements of a repeated field of a message. 
}
\section{Methods}{
\describe{
\item{\code{signature(object = "Message")}}{swap elements of a repeated field of a message}
}}
\references{
	See the \code{SwapElements} of the \code{Reflection} class, part of the protobuf library. 
	\url{http://code.google.com/apis/protocolbuffers/docs/reference/cpp/google.protobuf.message.html}
}
\keyword{methods}

