\name{ArrayInputStream-methods}
\docType{methods}
\alias{ArrayInputStream}
\alias{ArrayInputStream-methods}
\alias{ArrayInputStream,raw-method}
\title{Creates an ArrayInputStream}
\description{
Constructor for \linkS4class{ArrayInputStream} objects
}
\section{Methods}{
\describe{
\item{\code{signature(payload = "raw")}}{Creates a \linkS4class{ArrayInputStream} using the raw vector as the payload of the stream}
}}
\keyword{methods}

