\name{descriptor-methods}
\docType{methods}

\alias{descriptor}
\alias{descriptor-methods}
\alias{descriptor,Message-method}
\title{Get the descriptor of a message}
\description{
	Get the \linkS4class{Descriptor} associated
	with a \linkS4class{Message}
}
\section{Methods}{
\describe{
\item{\code{signature(object = "Message")}}{ Get the descriptor of the message, as a \linkS4class{Descriptor} instance }
}}
\keyword{methods}

