\name{fileDescriptor-methods}
\docType{methods}
\alias{fileDescriptor}
\alias{fileDescriptor-methods}
\alias{fileDescriptor,protobufDescriptor-method}
\alias{fileDescriptor,protobufMessage-method}
\title{gets the file descriptor of an object}
\description{
Gets the file descriptor of an object
}
\section{Methods}{
\describe{
\item{\code{signature(object = "protobufDescriptor")}}{ retrieves the file descriptor associated with this descriptor}
\item{\code{signature(object = "protobufMessage")}}{ retrieves the file descriptor associated with the descriptor of this message }
}}
\keyword{methods}

