\name{invokeLocally-methods}
\docType{methods}
\alias{invokeLocally}
\alias{invokeLocally-methods}
\alias{invokeLocally,MethodDescriptor,Message-method}
\title{invoke a protobuf rpc method locally}
\description{
invoke a protobuf rpc method locally
}
\section{Methods}{
\describe{
\item{\code{signature(method = "MethodDescriptor", message = "Message")}}{invoke a protobuf rpc method locally}
}}
\keyword{methods}
