\name{name}
\docType{methods}
\alias{name}
\alias{name-methods}
\alias{name,protobufDescriptor-method}
\alias{name,protobufFieldDescriptor-method}
\alias{name,protobufEnumDescriptor-method}
\alias{name,protobufServiceDescriptor-method}
\alias{name,protobufMethodDescriptor-method}

\title{Name or full name of a descriptor}
\description{
name or full name of a descriptor
}
\section{Methods}{
\describe{
	\item{\code{signature(object = "protobufDescriptor")}}{ ... }
	\item{\code{signature(object = "protobufFieldDescriptor")}}{ ... }
	\item{\code{signature(object = "protobufEnumDescriptor")}}{ ... }
	\item{\code{signature(object = "protobufServiceDescriptor")}}{ ... }
	\item{\code{signature(object = "protobufMethodDescriptor")}}{ ... }
}}
\keyword{methods}

