\name{name}
\docType{methods}
\alias{name}
\alias{name-methods}
\alias{name,protobufDescriptor-method}
\alias{name,protobufFieldDescriptor-method}
\title{Name or full name of a descriptor}
\description{
name or full name of a descriptor
}
\section{Methods}{
\describe{
	\item{\code{signature(object = "protobufDescriptor")}}{ the name or the full name of the message type descriptor}
	\item{\code{signature(object = "protobufFieldDescriptor")}}{ the name or the full name of the field descriptor}
}}
\keyword{methods}

