\name{nested_type-methods}
\alias{nested_type}
\alias{nested_type-methods}
\title{Extract a message type descriptor for a nested type}
\description{
	Extract a \linkS4class{Descriptor} nested in another 
	\linkS4class{Descriptor}
}
\seealso{
The method is implemented for the \linkS4class{Descriptor} class
}
\keyword{methods}

