\name{type-methods}

\alias{type}
\alias{type-methods}

\alias{cpp_type}
\alias{cpp_type-methods}

\alias{TYPE_DOUBLE}
\alias{TYPE_FLOAT}
\alias{TYPE_INT64}
\alias{TYPE_UINT64}
\alias{TYPE_INT32}
\alias{TYPE_FIXED64}
\alias{TYPE_FIXED32}
\alias{TYPE_BOOL}
\alias{TYPE_STRING}
\alias{TYPE_GROUP}
\alias{TYPE_MESSAGE}
\alias{TYPE_BYTES}
\alias{TYPE_UINT32}
\alias{TYPE_ENUM}
\alias{TYPE_SFIXED32}
\alias{TYPE_SFIXED64}
\alias{TYPE_SINT32}
\alias{TYPE_SINT64}

\alias{CPPTYPE_INT32}
\alias{CPPTYPE_INT64}
\alias{CPPTYPE_UINT32}
\alias{CPPTYPE_UINT64}
\alias{CPPTYPE_DOUBLE}
\alias{CPPTYPE_FLOAT}
\alias{CPPTYPE_BOOL}
\alias{CPPTYPE_ENUM}
\alias{CPPTYPE_STRING}
\alias{CPPTYPE_MESSAGE}

\title{Gets the type or the c++ type of a field}
\description{
Gets the type or the c++ type of a field
}
\seealso{
The method is implemented for the \linkS4class{protobufFieldDescriptor} class
}
\keyword{methods}

